Update reaction.py
Modifies the reaction.py to explicitly check if user provides Ea for a given reaction, when writing thermo YAML file for OpenMKM. Current behavior is wrong, user provided Ea is overridden by the `get_G_act` method.
Unsure why the comment in line 154 is different. I did not change this comment and the production version of pMuTT uses description identical to what I am using and matches defaults for the method.